requirements-dev.txt
-r requirements.txt
pytest>=9.0.3
pytest-mock>=3.14.0
pytest-asyncio>=0.24.0
ruff>=0.14.0
mypy>=1.11.0
lxml-stubs>=0.5.0
bandit>=1.7.10
pip-audit>=2.8.0
pytest-cov>=5.0.0
